refactor: TaskDetails를 SubjectContents로 이름 변경
