+++ src/components/calendar/SubjectContents/index.ts
@@ -1,0 +1,3 @@
+import SubjectContents from './SubjectContents';
+
+export default SubjectContents;
