--- src/components/calendar/TaskDetails/index.ts
@@ -1,3 +1,0 @@
-import TaskDetails from './TaskDetails';
-
-export default TaskDetails;
